Travel control added. Code from main activity moved to maps fragment
